Move atomic group helpers to common
