+++ miles/backends/megatron_utils/update_weight/update_weight_from_distributed/mixin.py
@@ -9,10 +9,11 @@ from miles.backends.training_utils.parallel import get_parallel_state
 from miles.utils.distributed_utils import get_gloo_group
 from miles.utils.timer import timer
 
-from ...megatron_to_hf import convert_to_hf, get_atomic_update_groups
+from ...megatron_to_hf import convert_to_hf
 from ..common import (
     all_gather_param,
     collect_named_tensors_for_weight_transfer,
+    get_atomic_update_groups,
     get_named_value_update_units,
     post_process_weights,
 )
